requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
azure-eventhub==5.11.4
azure-identity==1.15.0
azure-storage-file-datalake==12.14.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.2.0
structlog==23.2.0
prometheus-client==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.8.0
